Bug 1381542 - fixed a bug on `once`, added predicate and object's event handlers; r=ochameau
The contextual object on `once` wasn't set properly, added also unit test to
check this scenario.
The method `once` now can accept a a predicate's function, useful especially in
tests, when used with the returned promised.
The methods `on`, `once` (and `off`) can now accept an object with a method
`EventEmitter.handler` as DOM events does with `handleEvent`. That helps if
a different contextual object needs to be used, and / or to define a master
switch.