Bug 1381542 - refactored EventEmitter with old SDK event/core capabilities; r=ochameau
`EventEmitter` now can be used in the same way `sdk/event/core`, it also
integrates some of the SDK capabilities.
In addition the `emit` method is now unified in functionality, and is more
consistent with the nodejs counterpart (the listener doesn't receive the event

The method `once` now can accept a a predicate's function, useful especially in
tests, when used with the returned promised.
The methods `on`, `once` (and `off`) can now accept an object with a method
`EventEmitter.handler` as DOM events does with `handleEvent`. That helps if
a different contextual object needs to be used, and / or to define a master
switch.

MozReview-Commit-ID: DNxnjwmcC4W